Fix remaining clippy warnings across workspace
- Simplify let binding in cam-binder isp_session.rs
- Remove unnecessary usize casts in cam-binder metadata.rs
- Add Default impls for HeapAllocator, DebugService, ApplicationHolder
- Clean up impl block structure in cam-core lib.rs

All clippy warnings eliminated.

+++ cam-rust/cam-binder/src/metadata.rs
@@ -315,7 +315,7 @@ impl CameraMetadata {
                 _ => return Err(format!("unknown type: {}", data_type)),
             };
 
-            let value_start = header_size + entries_size as usize + value_offset as usize;
+            let value_start = header_size + entries_size + value_offset;
             let value_end = value_start + (count as usize * type_size);
             let value = data[value_start..value_end].to_vec();
